Cleaned up and added some logging

+++ deployer/src/main/java/no/f12/agiledeploy/deployer/ConfigurationServiceImpl.java
@@ -1,25 +1,29 @@
 package no.f12.agiledeploy.deployer;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.nio.channels.FileChannel;
 
+import org.apache.log4j.Logger;
 import org.springframework.stereotype.Component;
 
 @Component
 public class ConfigurationServiceImpl implements ConfigurationService {
 
+	private static final Logger LOG = Logger.getLogger(ConfigurationServiceImpl.class);
+
 	@Override
 	public void configure(File unpackDir, String environment) {
 		File propDir = new File(unpackDir, "properties/" + environment);
+		LOG.info("Updating configuration");
 		if (propDir.exists()) {
 			File[] files = propDir.listFiles();
 			for (File source : files) {
 				File target = new File(unpackDir, source.getName());
-				FileUtil.copyFile(source, target);
+				if (!target.exists()) {
+					FileUtil.copyFile(source, target);
+					LOG.info("Installed configuration file: " + source);
+				} else {
+					LOG.debug("File already exists, skipping " + source);
+				}
 			}
 
 		}
